Remove VERSION constant
Requiring the feed_parser.rb in the gemspec means all the dependencies
need to be installed to build the gem, which sucks.

It was also defined as a toplevel constant, which caused conflicts
with other libraries. Not easy to put in a module as there is already
a class called FeedParser, so we can't define a module with that name.

Easiest to just remove it and put the version in the gemspec by hand.